requirements.txt
asgiref==3.4.1
Django==3.0.14
django-taggit==1.2.0
Markdown==3.2
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
